backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
passlib[argon2]
python-jose[cryptography]
jinja2
python-multipart
